rmc_backend/reviews/serializers.py
from rest_framework import serializers #PROBLEM POINT
from .models import Course, Review

class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = '__all__'

class ReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = Review
        fields = '__all__'