rmc_backend/requirements.txt
django-cors-headers
djangorestframework
mysqlclient